css-ui-3_dev/html/caret-color-009.htm

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>CSS Basic User Interface Test: caret-color computed values</title>
<link href="mailto:florian@rivoal.net" rel="author" title="Florian Rivoal">
<link href="http://www.w3.org/TR/css3-ui/#caret-color" rel="help" title="7.2.1. Coloring the insertion caret: caret-color">
<link href="https://drafts.csswg.org/css-color/#currentcolor-color" rel="help">
<meta content="dom" name="flags">
<meta content="Test checks that auto and currentcolor compute to themselves for caret-color" name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#d1 {
	caret-color: auto;
}
#d2 {
	caret-color: currentcolor;
}
</style>
</head><body>
  <div id="d1"></div>
  <div id="d2"></div>
  <div id="log"></div>

<script>
test(
  function(){
   var d1 = document.getElementById("d1");
   assert_equals(window.getComputedStyle(d1)["caret-color"], "auto");
  }, "The computed value of auto should be auto");
test(
  function(){
   var d2 = document.getElementById("d2");
   assert_equals(window.getComputedStyle(d1)["caret-color"], "currentcolor");
  }, "The computed value of currentcolor should be currentcolor");
</script>


</body></html>
